// src/yenisi/src/uBkrVeriOperasyonu/interfaces/ICustomDefinedComplexServiceNode.ts
export interface ICustomDefinedComplexServiceNode {
    ActiveCd?: string;
    Data?: string;
    Description?: string;
    Guid?: string;
    Id?: string;
    IsCustomNode?: boolean;
    NodeTypeRef?: string;
    OperandTypeRef?: string;
    ParentGuidRef?: string;
    ParameterOrder?: number;
    ComplexMathExpression?: string;
}